Add function to execute multiple futures concurrently
To have a good performance the program needs to be able to fetch
multiple pages in parallel. The helper function takes a Vec of futures
returning a Vec<T>, executes them in parallel and flattens the Vecs into
one.
In a previous project I ran into timeout errors when allowing an
unlimited amount of parallelism. To not deal with the same issues here,
the helper function will execute at most <number of cores> futures in
parallel.