Implement /v1/crawl/{domain}
This function is the real meat of the service.
We start by fetching the user provided page. We will fetch every new
link we find in that result as long as it refers to the same domain.
Links which don't belong to that domain are just indexed but not
requested.
After the operation terminated, we tell the user how long it took.

The idiomatic thing to do would be to register this funcitonality only
with the POST method but because it's easier to test in the browser with
GET and because I don't know how third parties want to use the API I
decided to offer the same funcitonality also for the GET method.