Support building a docker image
To easily deploy the service with a cloud service provider, it would be
nice to have a quick way to create a docker image with it.

Because the docker container building the binary will not be the same
running it, we need to cross compile openssl. To do that I added it as a
dependency and enabled the "vendored" feature.

Building with docker will create a builder container which cross
compiles the binary for the musl library implementation (in order to use
it with the alpine image).

Building the binary is split into 2 steps.
1) Building all the dependencies
2) Building our actual program
This was done in order to reduce build times if the dependencies didn't
change between 2 builds.

To not run into troubles with SSL in the final docker image, we need to
also provide SSL certificates as well as the actual binary.